extract ContractProcessor from JobProcessor
