+++ worker/src/main/java/gov/cms/ab2d/worker/processor/ContractProcessor.java
@@ -1,0 +1,15 @@
+package gov.cms.ab2d.worker.processor;
+
+import gov.cms.ab2d.common.model.JobOutput;
+import gov.cms.ab2d.worker.processor.StreamHelperImpl.FileOutputType;
+import gov.cms.ab2d.worker.processor.domainmodel.ContractData;
+
+import java.io.FileNotFoundException;
+import java.nio.file.Path;
+import java.util.List;
+
+public interface ContractProcessor {
+
+    List<JobOutput> process(Path outputDirPath, ContractData contractData, FileOutputType contractType)
+            throws FileNotFoundException;
+}
